feat: add recycler view to search fragment

+++ app/src/main/java/com/example/weather/WeatherItem.kt
@@ -1,3 +1,3 @@
 package com.example.weather
 
-data class WeatherItem(val weather: String, val cityName: String, val temperature: Float)
+data class WeatherItem(val weather: String, val cityName: String, val temperature: Double)
